Tue, Oct 31, 2023, 3:10 PM +03:00

+++ credit/credit.c
@@ -37,7 +37,7 @@ int main(void)
         lengh ++;
     }
 
-    int cdnum = n;
+    long cdnum = n;
 
     for(int a = 0; a < lengh - 2; a++)
     {
